requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-standardized-errors==0.14.1
idna==3.10
inflection==0.5.1
ippanel==2.0.7
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
medianasms==1.1.0
pillow==10.4.0
psycopg2-binary==2.9.5
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
shortuuid==1.0.13
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
